requirements.txt
django
django-environ
djangorestframework
djangorestframework-xml
djangorestframework_simplejwt
gunicorn
uwsgi
whitenoise
dj-database-url
psycopg2-binary
six
numpy
opencv-python-headless
pytz
django-cors-headers
mysqlclient 
xgboost
scikit-learn